backend/requirements.txt
fastapi
uvicorn
pandas
numpy
fairlearn
scikit-learn
pydantic
python-multipart
google-generativeai
python-dotenv
firebase-admin
google-cloud-bigquery
